Fix gtk_message_dialog_new usage
Trying to build network-manager-openconnect on Ubuntu oneiric with
GTK3 (or GTK 2.24...), I've ran into a failure to build:

gtk_message_dialog_new () in auth-dialog/main.c has the build fail
because in at least GTK 2.24; this function expects a message format
along with the accompanying parameters rather than a preformatted
message. See the attached patch which fixes the issue.

[dwmw2: This looks like it was always a potential crash if the
        hostname or reason contained things like '%s'. Newer
        environments or versions of GTK+ actually make it fail to
        build, but it was always wrong.]