Show errors in auth-dialog properly
NM users weren't seeing the helpful warning messages that openconnect gave,
telling them that their SSL certificate was about to expire. That sucks;
fix it.

All messages are now rendered into the ssl_box immediately, and if the
obtain_cookie() call fails then the last icon is changed from a warning
to an error. Which is primitive, but until we fix openconnect to have
separate PRG_WARN and PRG_ERR log levels (which means an ABI change)
it's the best we can do. That *is* the difference between error and
warning anyway, fairly much.