Fix: Hitting cancel after failure causes the next attempt to abort immediately
If you cancel one connection while it's connecting, you get the 'Socket
connecton cancelled' error displayed as it gives up. Or a normal failure
should also suffice. The 'cancel' button is active at this point.
$DEITY knows why. It shouldn't be, because there's nothing to cancel.

If you hit it, a byte is written to the cancel_pipe. Then if you attempt
to connect to a new host, your connection attempt aborts immediately because
the cancel pipe is readable.

First, ensure the cancel button is marked as not sensitive after handling
an error return in cookie_obtained(). And also make sure we clear all bytes
from the cancel pipes before starting a connection, just in case.

Reported by Mike Miller.