service: If VPNGATEWAY is an IPv6 address, send it as such
If we get connected to the VPN via IPv6, pass that IPv6 address to the
daemon rather than just bailing out with an error.

With old NetworkManager, this will hit a g_return_if_fail(), and
result in no explicit route to the endpoint being configured, but
things may work anyway. It's at least not worse than the current
situation, where it's guaranteed to not work.