properties: Fix broken 'C' hotkey
'C' is assigned to both "_Cancel" (from the applet) and "_CA Certificate"
(from the VPN plugin).  The latter is also missing its mnemonic_widget
property, resulting in a Gtk-WARNING, but no tangible UI action, when the
user presses alt-C.

Assign 'E' to "CA C_ertificate" and point it to the CA certificate
chooser.