change multiprotocol support to use a new "protocol" option
The recently added Juniper VPN support was realized by adding a
service-type alias openconnect.nc. Thereby, the protocol was encoded in
the service-type. The service-type was traditionally the D-Bus bus name
of the VPN service and since multi-VPN it is a prefix for the bus name.
Another important role of the service-type is that it is stored inside a
connection, so that the vpn.service-type setting associates a connection
with its plugin.

The service-type should not be hacked to encode the openconnect protocol
option. That is, the service-type field already has different meanings,
and it should not be reused. Instead, add a new NM_OPENCONNECT_KEY_PROTOCOL
key.

However, in nm-connection-editor's UI we don't want to show the protocol
as an option inside the VPN dialog. Instead, the two protocols
"anyconnect" and "nc" should result in two separate entires in the
"add-connection" list. This way, the user first selects to create either
an "anyconnect" or "nc" VPN connection, and then the UI doesn't let
him switch protocol anymore.

This is realized by exposing the protocol as an "add-detail". When
populating the list of VPN types, connection-editor can ask the plugin
whether it supports multiple entires. Later, when adding the connection,
the protocol field is pre-filled with the "add-detail-key".

Note that the nm-openconnect plugin is mostly agnostic to the actual
protocol setting. When openconnect gains support for a new protocol, the
plugin's service part will just support it right away.
The only thing missing is to generate an additional add-connection entry
in the properties plugin. Note how the supported protocols are declared
in the .name file. So, the user could extend the plugin's .name file with
a new protocol and as a result, the new protocol would show up in the list.
However, then the plugin is unable to show a proper name and description
to the user, so to truly support new protocols, the plugin must be
extended too.
That also makes sense, as the .name file is not user-configuration. But
editing the .name file is a suitable workaround to support new protocols.

https://bugzilla.gnome.org/show_bug.cgi?id=767197