shared: add shared files
Add and "shared/" directory, that contains headers and source files
that can be used by all other components of the plugin (that is,
src/, properties/, and auth-dialog/).

This includes a "nm-default.h" header, which is supposed to be included
by (almost) every source file in the project as first. It provides
default includes (config.h, nm-glib.h).

Then there is shared/nm-utils/ directory, which contains utility
functions in form of header-only and source files. Those functions
are copied as-is from NetworkManager's shared/ repository and are
supposed not to be modified locally.