src: add logging macros to nm-openconnect-service
These macros follow the same pattern as used in other NetworkManager projects.
That is, they are enabled/disabled based on the logging configuration (for
now, only gl.debug).

Also, use g_print() instead of g_log(). The difference is, that a g_warning()
dumps core when running with G_DEBUG=fatal-warnings. g_warning() is thus
useful for assertion-like failures, but not for regular logging and not
to notify about possible runtime conditions.

Also, g_log() adds it's own prefixes that we don't want. Additionally,
by default g_message() writes to stdout, while g_warning() writes to
stderr. Overall, g_log() seems useful for logging from a library, but
not as logging mechanism for an application.

Also, have the logging macros prefix each line with "nm-openconnect[$PID]".
That is useful, because since multi-vpn-support, there may be multiple
openconnect services logging.